+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Handle edge-case where code block is not contained in a `p` tag.
+
 ## 0.1.1
 
 - Code blocks can be tagged with language syntax.